core, restapi: Introducing support for attaching disk snapshot
This patch adds the engine support for attaching a disk snapshot to a
VM (currently different VM than the one for which the VM
snapshot was taken).

Before this change, when attaching a disk to a vm only the active volume
was used, if the user wanted to see the disk content at some snapshot he
had to preview that snapshot.
After this change, a snapshot of a disk can be attached to another vm,
regardless of the disk not being marked as shareable - when doing so,
VDSM should create a temp snapshot allowing read/write access above the selected snapshot,
the above should happen when hotplugging a disk/ running a vm.

The following limitations currently applies on that ability:
1. The created temp snapshots is stored on the host local storage (the host that the vm is running on)
and not on the shared storage (domains) therefore the vm can't be migrated.
2. A disk snapshot can be attached to a different VM than the one of
which the snapshot (VM snapshot) was taken of.

Usage:
1. Navigate to the wanted disk snapshot from REST by accessing:
SERVER:PORT:/api/vms/GUID/snapshots/GUID/disks

2. POST the copied disk XML to:
http://SERVER:PORT/api/vms/GUID/disks/

Patch with lock handling is dependent on this one to ease the review.

OPEN ISSUES:
UI  - would be handled in a following patch (the
information is accessible through REST)

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/snapshots/SnapshotsManager.java
@@ -249,7 +249,7 @@ public class SnapshotsManager {
         VmDeviceUtils.setVmDevices(vm.getStaticData());
         ArrayList<DiskImage> images =
                 new ArrayList<DiskImage>(ImagesHandler.filterImageDisks(getDiskDao().getAllForVm(vm.getId()),
-                        false, true));
+                        false, true, true));
         for (DiskImage image : images) {
             image.setStorageIds(null);
         }
@@ -341,7 +341,7 @@ public class SnapshotsManager {
         if (!vmDevice.getDevice().equals(VmDeviceType.DISK.getName())) {
             return true;
         }
-        return getDiskDao().get(vmDevice.getDeviceId()).isAllowSnapshot();
+        return vmDevice.getSnapshotId() == null && getDiskDao().get(vmDevice.getDeviceId()).isAllowSnapshot();
     }
 
     /**
@@ -479,14 +479,14 @@ public class SnapshotsManager {
     }
 
     /**
-     * Remove all the disks which are allowed to be snapshot but not exist in the snapshot.
+     * Remove all the disks which are allowed to be snapshot but not exist in the snapshot and are not disk snapshots
      * @param vmId - The vm id which is being snapshot.
      * @param diskIdsFromSnapshot - An image group id list for images which are part of the VM.
      */
     private void removeDisksNotInSnapshot(Guid vmId, List<Guid> diskIdsFromSnapshot) {
         for (VmDevice vmDevice : getVmDeviceDao().getVmDeviceByVmIdTypeAndDevice(
                 vmId, VmDeviceGeneralType.DISK, VmDeviceType.DISK.getName())) {
-            if (!diskIdsFromSnapshot.contains(vmDevice.getDeviceId())) {
+            if (!diskIdsFromSnapshot.contains(vmDevice.getDeviceId()) && vmDevice.getSnapshotId() == null) {
                 Disk disk = getDiskDao().get(vmDevice.getDeviceId());
                 if (disk.isAllowSnapshot()) {
                     getBaseDiskDao().remove(vmDevice.getDeviceId());
